Added the changes for the deals first showing an error, removed the state from products to the corresponding pages, changed deals so it made the request for deals only from the backend
