--- src/components/pages/Deals.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-import Products from '../products/Products'
-
-const Deals = () => {
-    return <div className="row">
-        <Products deals />
-</div>
-}
-
-export default Deals
