+++ src/components/pages/deals/deals.css
@@ -1,0 +1,4 @@
+.noDeals {
+    color: red;
+    margin: auto;
+}
