+++ src/components/pages/home/Home.js
@@ -1,12 +1,24 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 
 import Products from '../../products/Products'
+import itemsServices from '../../../services/items'
+
 import './home.css'
 
-const home = () => {
+const Home = () => {
+    const [items, setItems] = useState([])
+
+    useEffect(() => {
+        const fetch_data = async () => {
+            const response_data = await itemsServices.getAllItems()
+            setItems(response_data.items)
+        }
+        fetch_data()        
+    }, [])
+
     return (
-        <Products />
+        <Products items={items}/>
     )
 }
 
-export default home
+export default Home
