+++ src/components/products/products.css
@@ -3,8 +3,3 @@
     box-sizing: border-box;
     align-items: center;
 }
-
-.noDeals {
-    color: red;
-    margin: auto;
-}
